requirements.txt
fastapi
uvicorn
python-dotenv
openai
sqlmodel
python-jose[cryptography]
passlib[bcrypt]
aiofiles
